requirements.txt
sentence-transformers
faiss-cpu
numpy
requests
nltk
pandas
matplotlib
seaborn
wordcloud
